A few routines
`coefficienti.m` is a MATLAB function which computes the coefficients of any 1D finite difference scheme (explicit as well as compact). It takes in input two arrays containing the coordinates of known and unknown values and two arrays containing the order of derivation of those knowns and unknonws (the last input is the collocation point of the scheme, _i.e._ the point where the coefficient 1 is set). Run `help coefficienti` for further details or `open coefficienti` to gain deeper knowledge.

`coefficienti_upgrade.m` is an extension of `coefficienti.m` and accepts also integral values, whose order of derivation is `-1`. Each integral value should be inserted in both coordinate and order of derivation arrays (`[... 2 2.5 ...]` and `[... -1 -1 ...]` are relative to an integral from 2 to 2.5. Type `open coefficienti_upgrade` to see the code.

`errtrunc.m` is a "library" which computes the truncation error for several finite difference schemes (explicit as well as compact).