Bash script to set up the screens
The bash script (which makes use of a sed script), hardcodes an
empirical rule to decide which screen should be turned on/off.

This logic should be factored out in another script that the user is
free to change at will, provided it parses the output of the command
`xrandr | availscreens.sed` and gives an output that the rest of this
script can parse and transform in a xrandr command that sets thing up.